Jwt authenticated routes

+++ app/router/survey.py
@@ -1,7 +1,12 @@
+from http import HTTPStatus
 from fastapi import APIRouter, HTTPException
 from typing import List
+from service.user import User
 from model.survey import SurveyBase, SurveyCreate, SurveyUpdate, SurveyResponse
+from model.user import UserType
 from service.survey import Survey as SurveyService
+from internal.security import get_current_user, verify_permissions
+from fastapi import APIRouter, Depends, HTTPException
 
 router = APIRouter()
 
@@ -11,7 +16,11 @@ router = APIRouter()
     summary="Create a new survey", 
     description="This endpoint allows you to create a new survey."
 )
-async def create_survey(survey: SurveyCreate):
+async def create_survey(survey: SurveyCreate,
+    current_user: User = Depends(get_current_user)
+):
+    verify_permissions(current_user, UserType.orgadmin, {'orgid': survey.orgid})
+
     return await SurveyService.create_survey(survey)
 
 @router.get(
@@ -20,7 +29,11 @@ async def create_survey(survey: SurveyCreate):
     summary="Retrieve all surveys", 
     description="Retrieve a list of all surveys available."
 )
-async def get_surveys():
+async def get_surveys(
+    current_user: User = Depends(get_current_user)
+):
+    verify_permissions(current_user, UserType.superadmin)
+
     surveys = await SurveyService.get_all_surveys()
     return surveys
 
@@ -30,10 +43,13 @@ async def get_surveys():
     summary="Retrieve a survey by ID", 
     description="Retrieve a specific survey by its ID."
 )
-async def get_survey(surveyid: int):
+async def get_survey(surveyid: int,
+    current_user: User = Depends(get_current_user)
+):
     survey = await SurveyService.get_survey(surveyid)
-    if survey is None:
-        raise HTTPException(status_code=404, detail="Survey not found")
+    
+    verify_permissions(current_user, UserType.employee, {'orgid': survey.orgid})
+    
     return survey
 
 @router.get(
@@ -42,7 +58,11 @@ async def get_survey(surveyid: int):
     summary="Retrieve all surveys by organization ID",
     description="Retrieve a list of all surveys associated with a specific organization ID."
 )
-async def get_surveys_by_org(orgid: int):
+async def get_surveys_by_org(orgid: int,
+    current_user: User = Depends(get_current_user)
+):
+    verify_permissions(current_user, UserType.employee, {'orgid': orgid})
+    
     surveys = await SurveyService.get_survey_by_org(orgid)
     return surveys
 
@@ -52,10 +72,14 @@ async def get_surveys_by_org(orgid: int):
     summary="Update a survey", 
     description="Update the details of a specific survey by its ID."
 )
-async def update_survey(surveyid: int, survey: SurveyUpdate):
+async def update_survey(surveyid: int, survey: SurveyUpdate,
+    current_user: User = Depends(get_current_user)
+):
+    verify_permissions(current_user, UserType.orgadmin, {'orgid': survey.orgid})
+
     result = await SurveyService.update_survey(surveyid, survey)
     if not result:
-        raise HTTPException(status_code=404, detail="Survey not found")
+        raise HTTPException(HTTPStatus.NOT_FOUND, 'Survey not found')
     return result
 
 @router.delete(
@@ -63,19 +87,28 @@ async def update_survey(surveyid: int, survey: SurveyUpdate):
     summary="Delete a survey", 
     description="Delete a specific survey by its ID."
 )
-async def delete_survey(surveyid: int):
+async def delete_survey(surveyid: int,
+    current_user: User = Depends(get_current_user)
+):
+    survey = await SurveyService.get_survey(surveyid)
+    
+    verify_permissions(current_user, UserType.orgadmin, {'orgid': survey.orgid})
+
     result = await SurveyService.delete_survey(surveyid)
     if not result:
-        raise HTTPException(status_code=404, detail="Survey not found")
+        raise HTTPException(HTTPStatus.NOT_FOUND, 'Survey not found')
     return {"message": "Survey deleted successfully"}
 
-# rota para pesquisar as surveys não respondidas pelo usuario
 @router.get(
     "/unresponded/{employee_id}",
     response_model=List[SurveyResponse],
     summary="Retrieve all unresponded surveys",
     description="Retrieve a list of all surveys that have not been responded by the user."
 )
-async def get_unresponded_surveys(employee_id: int):
+async def get_unresponded_surveys(employee_id: int,
+    current_user: User = Depends(get_current_user)
+):
+    verify_permissions(current_user, UserType.employee, {'orgid': employee_id})
+
     surveys = await SurveyService.get_unresponded_surveys(employee_id)
     return surveys
